backend/oslomet_app/requirements.txt
fastapi
pydantic
pydantic-settings
aiomysql
pymysql
aiohttp
uvicorn
python-dotenv
tqdm
gspread
gspread-formatting
oauth2client
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
pandas
mysql-connector-python
python-multipart
passlib[bcrypt]
